--- src/stories/components/svg/dash-board.tsx
@@ -1,30 +1,0 @@
-import * as React from 'react';
-
-interface Props {
-    width?: number
-    height?: number
-    onClick?: () => void
-    fill?: string
-}
-
-const Dashboard = ({ width = 130, height = 26, fill = '#211634', ...props }: Props) => {
-  return (
-        <svg
-            width={width}
-            height={height}
-            viewBox="0 0 130 26"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-            {...props}
-        >
-            <path
-                fillRule="evenodd"
-                clipRule="evenodd"
-                d="M6 0a6 6 0 00-6 6v14a6 6 0 006 6h118a6 6 0 006-6V6a6 6 0 00-6-6H6zm36.7 18c-1.927 0-3.914-.884-3.914-3.238h1.824c0 1.141.815 1.754 2.09 1.754 1.334 0 2.06-.527 2.06-1.34 0-.91-1.107-1.229-2.349-1.587-1.584-.457-3.387-.977-3.387-2.864C39.024 9.027 40.758 8 42.73 8c2.016 0 3.691 1.056 3.691 3.153h-1.853c0-.985-.578-1.67-1.838-1.67-1.23 0-1.868.557-1.868 1.242 0 .876 1.088 1.176 2.323 1.516 1.618.445 3.488.96 3.488 2.934 0 1.94-2.105 2.825-3.973 2.825zM16 17.843v-9.7h3.054c3.677 0 5.307 2.182 5.307 4.85 0 2.668-1.63 4.85-5.307 4.85H16zm1.705-8.288v6.89h1.349c2.565 0 3.454-1.626 3.454-3.452s-.89-3.438-3.454-3.438h-1.35zm17.036 8.288h1.912l-3.81-9.686H30.59l-3.81 9.686h1.912l1.157-3.11h3.736l1.156 3.11zm-4.373-4.508l1.349-3.666 1.349 3.666h-2.698zm18.886 4.508h1.705v-4.28h4.477v4.28h1.705V8.157h-1.705v3.994H50.96V8.143h-1.705v9.7zm10.921-9.686v9.686h3.855c2.787 0 3.913-1.241 3.913-2.625 0-1.112-.652-2.097-1.971-2.482 1.023-.414 1.527-1.141 1.527-2.111 0-1.327-1.127-2.468-3.47-2.468h-3.854zm3.855 3.994h-2.15V9.555h2.15c1.304 0 1.704.613 1.704 1.284 0 .613-.4 1.312-1.704 1.312zm0 4.294h-2.15v-2.882h2.15c1.556 0 2.075.685 2.075 1.427s-.519 1.455-2.075 1.455zm6.628-3.395c0 2.667 1.63 4.95 4.848 4.95 3.217 0 4.847-2.283 4.847-4.95 0-2.668-1.63-4.936-4.847-4.936-3.217 0-4.848 2.268-4.848 4.936zm1.838 0c0-1.812.89-3.41 3.01-3.41 2.105 0 3.01 1.598 3.01 3.41 0 1.826-.905 3.424-3.01 3.424-2.12 0-3.01-1.598-3.01-3.424zm20.15 4.793h-1.912l-1.156-3.11h-3.736l-1.156 3.11h-1.913l3.81-9.686h2.254l3.81 9.686zM87.71 9.67l-1.349 3.666h2.698L87.71 9.67zm15.283 1.412c0-1.669-1.112-2.924-3.75-2.924h-4.018v9.686h1.705V14.02h1.349l2.653 3.823h2.002l-2.698-3.894c1.927-.271 2.757-1.413 2.757-2.868zM96.93 9.57h2.194c1.512 0 2.031.713 2.031 1.512 0 .813-.519 1.527-2.031 1.527H96.93V9.569zm11.763-1.426h-3.054v9.7h3.054c3.676 0 5.307-2.182 5.307-4.85 0-2.668-1.631-4.85-5.307-4.85zm-1.349 8.302v-6.89h1.349c2.564 0 3.454 1.612 3.454 3.438s-.89 3.452-3.454 3.452h-1.349z"
-                fill={fill}
-            />
-        </svg>
-  );
-};
-
-export default Dashboard;
