+++ src/stories/components/svg/expenses.tsx
@@ -1,0 +1,29 @@
+import * as React from 'react';
+
+interface Props {
+    width?: number
+    height?: number
+    fill?: string
+}
+
+const EXPENSES = ({ width = 130, height = 26, fill = '#211634', ...props }: Props) => {
+  return (
+    <svg
+    width={width}
+    height={height}
+    viewBox="0 0 130 26"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M6 0a6 6 0 00-6 6v14a6 6 0 006 6h118a6 6 0 006-6V6a6 6 0 00-6-6H6zm19.864 19H18.28V8.12h7.584v1.776H20.36v2.624h5.024v1.776H20.36v2.928h5.504V19zm11-10.88l2.544.032-3.856 5.44 3.856 5.44-2.56-.032-2.56-3.6-2.576 3.632-2.56-.032 3.84-5.424-3.84-5.456 2.544.032 2.576 3.616 2.592-3.648zm6.323.016V19h2.08v-3.568h2.128c3.088 0 4.336-1.728 4.336-3.648 0-1.936-1.248-3.648-4.336-3.648h-4.208zm4.048 5.504h-1.968V9.912h1.968c1.744 0 2.24 1.04 2.24 1.872 0 .816-.496 1.856-2.24 1.856zM62.755 19H55.17V8.12h7.584v1.776H57.25v2.624h5.024v1.776H57.25v2.928h5.504V19zm5.527 0h-2.08V8.136h2.56l4.704 7.904V8.12h2.08V19h-2.56l-4.704-7.76V19zm14.977.176c-2.096 0-4.256-.992-4.256-3.648h2.224c0 1.152.784 1.76 2.032 1.76 1.312 0 2.016-.512 2.016-1.36 0-.922-1.113-1.25-2.382-1.622-1.68-.494-3.634-1.068-3.634-3.21 0-1.904 1.856-3.136 4.032-3.136 2.256 0 4.064 1.232 4.064 3.616H85.1c0-1.04-.528-1.744-1.808-1.744-1.216 0-1.808.592-1.808 1.264 0 .869 1.087 1.17 2.344 1.52 1.73.48 3.784 1.052 3.784 3.312 0 2.256-2.336 3.248-4.352 3.248zM98.63 19h-7.583V8.12h7.584v1.776h-5.504v2.624h5.024v1.776h-5.024v2.928h5.504V19zm7.224.176c-2.096 0-4.256-.992-4.256-3.648h2.224c0 1.152.784 1.76 2.032 1.76 1.312 0 2.016-.512 2.016-1.36 0-.922-1.113-1.25-2.382-1.622-1.681-.494-3.634-1.068-3.634-3.21 0-1.904 1.856-3.136 4.032-3.136 2.256 0 4.064 1.232 4.064 3.616h-2.256c0-1.04-.528-1.744-1.808-1.744-1.216 0-1.808.592-1.808 1.264 0 .869 1.086 1.17 2.344 1.52 1.73.48 3.784 1.052 3.784 3.312 0 2.256-2.336 3.248-4.352 3.248z"
+      fill={fill}
+    />
+  </svg>
+  );
+};
+
+export default EXPENSES;
