+++ src/stories/components/svg/toggle-dark.tsx
@@ -1,32 +1,25 @@
 import * as React from 'react';
 
 interface Props {
-
     width?: number;
     height?: number;
     fill?: string
-    iconColor?: string
 }
-const ToggleDarkMode = ({ fill = '#31424E', iconColor = '#EDEFFF', height = 33, width = 33, ...props }: Props) => {
+const ToggleDarkMode = ({ fill = '#31424E', height = 32, width = 32, ...props }: Props) => {
   return (
         <svg
             width={width}
             height={height}
-            viewBox="0 0 33 33"
+            viewBox="0 0 32 32"
             fill="none"
             xmlns="http://www.w3.org/2000/svg"
             {...props}
         >
             <path
-                d="M.003 6.498a6 6 0 016-6h20.002a6 6 0 016 6V26.5a6 6 0 01-6 6H6.003a6 6 0 01-6-6V6.498z"
+                d="M16.06 11.173a4.81 4.81 0 00-2.649.79 4.718 4.718 0 00-1.758 2.109 4.647 4.647 0 00-.274 2.718 4.685 4.685 0 001.302 2.41 4.784 4.784 0 002.44 1.29c.924.182 1.882.09 2.753-.267a4.752 4.752 0 002.14-1.732 4.66 4.66 0 00.803-2.614 4.681 4.681 0 00-1.395-3.32 4.804 4.804 0 00-3.361-1.384zm0 7.853a3.243 3.243 0 01-1.785-.535 3.182 3.182 0 01-1.184-1.424 3.135 3.135 0 01-.183-1.833c.125-.615.43-1.18.88-1.624a3.227 3.227 0 011.646-.869 3.252 3.252 0 011.857.181c.587.24 1.089.647 1.442 1.169a3.143 3.143 0 01-.4 4.006 3.249 3.249 0 01-2.273.93zm-.803-10.169V6.97a.783.783 0 01.155-.67.801.801 0 01.629-.299.813.813 0 01.628.299.789.789 0 01.155.67v1.904a.792.792 0 01-.282.444.81.81 0 01-1.003 0 .792.792 0 01-.282-.444v-.016zm5.32 2.642a.788.788 0 010-1.095l1.357-1.34a.799.799 0 01.581-.354.813.813 0 01.647.215.792.792 0 01.246.629.786.786 0 01-.333.588L21.717 11.5a.809.809 0 01-.554.23.764.764 0 01-.587-.23zm5.39 4.497c0 .21-.084.412-.234.56a.808.808 0 01-.569.233h-1.872a.812.812 0 01-.902-.429.784.784 0 01.223-.964.807.807 0 01.68-.154h1.92a.81.81 0 01.544.224c.146.14.233.33.243.53h-.032zm-17.17.793h-1.88a.813.813 0 01-.64-.186.794.794 0 01-.277-.599.785.785 0 01.277-.6.807.807 0 01.64-.185h1.928c.178.04.336.138.45.278a.787.787 0 010 .99.805.805 0 01-.45.278l-.048.024zm8.035 6.346v1.896a.785.785 0 01-.435.89.812.812 0 01-.976-.22.79.79 0 01-.156-.67v-1.896a.783.783 0 01.155-.671.802.802 0 01.629-.299.812.812 0 01.628.299.79.79 0 01.155.67zm-5.327-2.681a.788.788 0 010 1.094l-1.374 1.373a.809.809 0 01-.555.23.763.763 0 01-.554-.23.788.788 0 010-1.095l1.374-1.372a.81.81 0 011.109 0zM9.038 10.142a.784.784 0 01.134-.946.807.807 0 01.959-.133l1.374 1.341a.788.788 0 010 1.095.756.756 0 01-.57.23.809.809 0 01-.555-.23l-1.342-1.357zm14.037 11.684a.788.788 0 010 1.095.757.757 0 01-.554.23.809.809 0 01-.555-.23l-1.39-1.34a.791.791 0 010-1.127.812.812 0 011.141 0l1.358 1.372z"
                 fill={fill}
             />
-            <path
-                d="M16.065 11.672a4.81 4.81 0 00-2.65.789 4.718 4.718 0 00-1.758 2.11 4.647 4.647 0 00-.275 2.718 4.685 4.685 0 001.302 2.41 4.784 4.784 0 002.44 1.29c.925.182 1.883.09 2.754-.267a4.753 4.753 0 002.14-1.732 4.661 4.661 0 00.804-2.614 4.681 4.681 0 00-1.395-3.321 4.805 4.805 0 00-3.362-1.383zm0 7.853a3.244 3.244 0 01-1.786-.535 3.183 3.183 0 01-1.184-1.424 3.135 3.135 0 01-.183-1.833 3.16 3.16 0 01.88-1.625 3.228 3.228 0 011.646-.868 3.253 3.253 0 011.857.18c.587.24 1.089.648 1.442 1.17a3.144 3.144 0 01-.4 4.006 3.25 3.25 0 01-2.273.93zm-.804-10.17V7.469a.783.783 0 01.155-.671.802.802 0 01.628-.299.812.812 0 01.629.299.789.789 0 01.155.67v1.905a.792.792 0 01-.282.443.81.81 0 01-1.003 0 .792.792 0 01-.282-.443v-.016zm5.32 2.642a.788.788 0 010-1.094l1.358-1.341a.8.8 0 01.58-.354.813.813 0 01.648.216.792.792 0 01.245.628.786.786 0 01-.332.589l-1.358 1.356a.809.809 0 01-.555.23.763.763 0 01-.587-.23zm5.391 4.498c0 .21-.084.412-.235.56a.808.808 0 01-.568.233h-1.872a.812.812 0 01-.903-.43.784.784 0 01.223-.963.808.808 0 01.68-.154h1.92a.81.81 0 01.544.224c.146.14.233.33.243.53h-.032zm-17.171.793H6.92a.813.813 0 01-.64-.186.794.794 0 01-.278-.6.785.785 0 01.277-.598.808.808 0 01.64-.187H8.85c.177.04.336.139.45.28a.787.787 0 010 .989.805.805 0 01-.45.278l-.048.024zm8.035 6.346v1.896a.783.783 0 01-.155.67.801.801 0 01-.628.3.813.813 0 01-.629-.3.79.79 0 01-.155-.67v-1.896a.783.783 0 01.155-.67.802.802 0 01.629-.3.812.812 0 01.628.3.788.788 0 01.155.67zm-5.327-2.681a.788.788 0 010 1.095l-1.375 1.372a.809.809 0 01-.554.23.763.763 0 01-.554-.23.788.788 0 010-1.095l1.374-1.372a.81.81 0 011.108 0zM9.041 10.64a.785.785 0 01.134-.947.807.807 0 01.958-.132l1.374 1.34a.788.788 0 010 1.095.754.754 0 01-.57.23.809.809 0 01-.554-.23l-1.342-1.356zm14.037 11.684a.788.788 0 010 1.095.755.755 0 01-.554.23.809.809 0 01-.554-.23l-1.39-1.34a.791.791 0 010-1.127.812.812 0 011.14 0l1.358 1.372z"
-                fill={iconColor}
-            />
-        </svg>
-  );
+        </svg>);
 };
 
 export default ToggleDarkMode;
